fix(studio assignments): restore Playbooks/Missions flip on hub
I overcorrected on HUB-A — dropped the tab system the classic page
had and replaced it with three vertical CD teaser strips
(Most-used playbooks · Marketplace recommendations). The user wants
to flip between Playbooks and Missions in place, not see one row of
three teasers.

Hub layout now:
- Editorial top
- Start something (EntryGrid)
- Needs approval cards (conditional)
- Flip tabs: Playbooks · Missions (URL-synced via ?tab=)
- Tab body — full PlaybooksBody / MissionsBody

To share the body content between hub tabs and standalone routes,
extracted PlaybooksBody and MissionsBody. Standalone /playbooks and
/missions pages now wrap those bodies with their own editorial top
and create CTAs.
