+++ frontend/components/assignments/studio/studio-playbooks-page.tsx
@@ -1,132 +1,40 @@
 'use client'
 
 /**
- * Studio Playbooks library — CD round 4 standalone Playbooks page.
- *
- * Featured strip on top (MktCard-shaped picks from the same data),
- * then the full library as a grid of PlaybookCards or a flat table.
- * Scope tabs (All / Mine / Workspace / Imported) layer over a search
- * box and sort menu.
- *
- * "Mine" / "Workspace" / "Imported" use the playbook fields when
- * available; missing fields fall through to "All".
+ * Studio Playbooks library (standalone /playbooks route).
+ * Wraps PlaybooksBody with the editorial top and create CTAs.
  */
 
-import { useMemo, useState } from 'react'
+import { useState } from 'react'
 import { useRouter } from 'next/navigation'
 import dynamic from 'next/dynamic'
-import {
-  LayoutGrid,
-  List as ListIcon,
-  Store,
-  Upload,
-  Plus,
-  Search,
-  ArrowDownNarrowWide,
-  BookMarked,
-  Clock,
-  Zap as TriggerZap,
-  Hand,
-} from 'lucide-react'
+import { Store, Upload, Plus } from 'lucide-react'
 
 import { useWorkflowPlaybooks } from '@/hooks/use-playbook-api'
 import { useMissions } from '@/hooks/use-missions-api'
 
-import { StatusHead } from './status-head'
-import { PlaybookCard, type PlaybookLike } from './playbook-card'
-
-type View = 'grid' | 'list'
-type Scope = 'all' | 'mine' | 'workspace' | 'imported'
-type Sort = 'most_run' | 'recent' | 'success'
+import { PlaybooksBody } from './playbooks-body'
 
 const CreatePlaybookModal = dynamic(
   () => import('@/components/workflows/create-playbook-modal').then((m) => m.CreatePlaybookModal),
   { ssr: false },
 )
 
-function trigger(p: PlaybookLike): { label: string; kind: 'cron' | 'webhook' | 'manual' } {
-  const t = (p.trigger_type || '').toLowerCase()
-  if (t === 'cron' || p.schedule_config?.cron) {
-    return { label: p.schedule_config?.cron || 'scheduled', kind: 'cron' }
-  }
-  if (t === 'webhook') return { label: 'webhook', kind: 'webhook' }
-  return { label: 'manual', kind: 'manual' }
-}
-
 export function StudioPlaybooksPage() {
   const router = useRouter()
-
-  const [view, setView] = useState<View>('grid')
-  const [scope, setScope] = useState<Scope>('all')
-  const [sort, setSort] = useState<Sort>('most_run')
-  const [search, setSearch] = useState('')
   const [playbookOpen, setPlaybookOpen] = useState(false)
 
-  const { data, isLoading } = useWorkflowPlaybooks({ limit: 100 })
+  const { data } = useWorkflowPlaybooks({ limit: 1 })
   const { data: missionData } = useMissions({ limit: 1 })
 
-  const all = (data as { items?: PlaybookLike[] } | undefined)?.items ?? []
+  const totalPlaybooks =
+    (data as { total?: number; items?: unknown[] } | undefined)?.total ??
+    (data as { items?: unknown[] } | undefined)?.items?.length ??
+    0
   const missionCount = missionData?.total ?? missionData?.missions?.length ?? 0
-
-  const counts = useMemo(() => {
-    const mine = all.filter((p) => (p as { is_mine?: boolean }).is_mine).length
-    const workspace = all.filter(
-      (p) => (p as { is_public?: boolean }).is_public === false,
-    ).length
-    const imported = all.filter(
-      (p) => (p as { source?: string }).source === 'marketplace',
-    ).length
-    return { all: all.length, mine, workspace, imported }
-  }, [all])
-
-  const filtered = useMemo(() => {
-    let list = all
-    if (scope === 'mine') {
-      list = list.filter((p) => (p as { is_mine?: boolean }).is_mine)
-    } else if (scope === 'workspace') {
-      list = list.filter(
-        (p) => (p as { is_public?: boolean }).is_public === false,
-      )
-    } else if (scope === 'imported') {
-      list = list.filter(
-        (p) => (p as { source?: string }).source === 'marketplace',
-      )
-    }
-    if (search.trim()) {
-      const q = search.trim().toLowerCase()
-      list = list.filter(
-        (p) =>
-          (p.name || '').toLowerCase().includes(q) ||
-          (p.description || '').toLowerCase().includes(q) ||
-          (p.category || '').toLowerCase().includes(q),
-      )
-    }
-    const sorted = [...list].sort((a, b) => {
-      if (sort === 'recent') {
-        const ta = (a as { updated_at?: string }).updated_at || ''
-        const tb = (b as { updated_at?: string }).updated_at || ''
-        return tb.localeCompare(ta)
-      }
-      if (sort === 'success') {
-        return (b.success_rate ?? 0) - (a.success_rate ?? 0)
-      }
-      return (b.use_count ?? 0) - (a.use_count ?? 0)
-    })
-    return sorted
-  }, [all, scope, sort, search])
-
-  const featured = useMemo(() => all.filter((p) => p.is_featured).slice(0, 5), [all])
-
-  const totalRuns = all.reduce((acc, p) => acc + (p.use_count ?? 0), 0)
-
-  const handleRun = (p: PlaybookLike) => {
-    const id = p.template_id || (p.id != null ? String(p.id) : '')
-    if (id) router.push(`/playbooks?id=${encodeURIComponent(id)}` as any)
-  }
 
   return (
     <div className="cc-page">
-      {/* Editorial top */}
       <div className="cc-headrow">
         <div className="cc-head">
           <p className="cc-eyebrow">The work · reusable routines</p>
@@ -160,11 +68,10 @@ export function StudioPlaybooksPage() {
         </div>
       </div>
 
-      {/* Sub-tabs */}
       <nav className="cc-tabs" aria-label="Assignments sections">
         <button type="button" className="cc-tab active" aria-current="page">
           <span>Playbooks</span>
-          {all.length > 0 && <span className="cc-tab-ct">{all.length}</span>}
+          {totalPlaybooks > 0 && <span className="cc-tab-ct">{totalPlaybooks}</span>}
         </button>
         <button
           type="button"
@@ -176,316 +83,12 @@ export function StudioPlaybooksPage() {
         </button>
       </nav>
 
-      {/* Toolbar */}
-      <div
-        style={{
-          display: 'flex',
-          alignItems: 'center',
-          gap: 10,
-          flexWrap: 'wrap',
-        }}
-      >
-        <div className="scope-tabs" role="group" aria-label="Scope">
-          {(
-            [
-              { k: 'all', l: 'All', c: counts.all },
-              { k: 'mine', l: 'Mine', c: counts.mine },
-              { k: 'workspace', l: 'Workspace', c: counts.workspace },
-              { k: 'imported', l: 'Imported', c: counts.imported },
-            ] as { k: Scope; l: string; c: number }[]
-          ).map((s) => (
-            <button
-              key={s.k}
-              type="button"
-              className={scope === s.k ? 'on' : ''}
-              onClick={() => setScope(s.k)}
-            >
-              {s.l}
-              <span className="ct">{s.c}</span>
-            </button>
-          ))}
-        </div>
-
-        <div className="view-toggle" role="group" aria-label="View">
-          <button
-            type="button"
-            className={view === 'grid' ? 'on' : ''}
-            onClick={() => setView('grid')}
-          >
-            <LayoutGrid style={{ width: 11, height: 11 }} />
-            Grid
-          </button>
-          <button
-            type="button"
-            className={view === 'list' ? 'on' : ''}
-            onClick={() => setView('list')}
-          >
-            <ListIcon style={{ width: 11, height: 11 }} />
-            List
-          </button>
-        </div>
-
-        <div style={{ marginLeft: 'auto', display: 'inline-flex', alignItems: 'center', gap: 8 }}>
-          <label
-            style={{
-              display: 'inline-flex',
-              alignItems: 'center',
-              gap: 6,
-              background: 'var(--paper)',
-              border: '1px solid hsl(var(--border))',
-              borderRadius: 6,
-              padding: '0 10px',
-              height: 28,
-              minWidth: 200,
-            }}
-          >
-            <Search style={{ width: 12, height: 12, color: 'var(--fg-3)' }} />
-            <input
-              type="text"
-              value={search}
-              onChange={(e) => setSearch(e.target.value)}
-              placeholder="Search playbooks…"
-              style={{
-                border: 0,
-                outline: 'none',
-                background: 'transparent',
-                fontSize: 12,
-                color: 'var(--fg)',
-                flex: 1,
-              }}
-            />
-          </label>
-          <button
-            type="button"
-            className="l-btn"
-            onClick={() =>
-              setSort((s) =>
-                s === 'most_run' ? 'recent' : s === 'recent' ? 'success' : 'most_run',
-              )
-            }
-            title="Toggle sort"
-          >
-            <ArrowDownNarrowWide style={{ width: 11, height: 11 }} />
-            Sort:{' '}
-            {sort === 'most_run'
-              ? 'most run'
-              : sort === 'recent'
-              ? 'recent'
-              : 'success'}
-          </button>
-        </div>
-      </div>
-
-      {/* Featured strip */}
-      {featured.length > 0 && (
-        <section>
-          <StatusHead
-            pip="hsl(82 50% 22%)"
-            label="★ Featured · proven in the wild"
-            count={`${featured.length} picks`}
-          />
-          <div className="mkt-strip">
-            {featured.map((p) => (
-              <button
-                key={String(p.template_id || p.id)}
-                type="button"
-                className="mkt-card"
-                onClick={() => handleRun(p)}
-              >
-                <div className="top">
-                  <span className="ic">
-                    <BookMarked style={{ width: 14, height: 14, strokeWidth: 1.6 }} />
-                  </span>
-                  <span className="nm">{p.name || 'Untitled'}</span>
-                </div>
-                <span className="tag">★ Featured</span>
-                {p.description && <div className="desc">{p.description}</div>}
-                <div className="foot">
-                  <span>{p.category || 'playbook'}</span>
-                  <span className="dl">{p.use_count ?? 0} runs</span>
-                </div>
-              </button>
-            ))}
-          </div>
-        </section>
-      )}
-
-      {/* Main grid/list */}
-      <section>
-        <StatusHead
-          pip="hsl(var(--foreground))"
-          label="Your library"
-          count={`${all.length} playbooks · ${totalRuns.toLocaleString()} total runs`}
-        />
-        {isLoading ? (
-          <div
-            className="pg-panel"
-            style={{
-              padding: 24,
-              textAlign: 'center',
-              color: 'var(--fg-3)',
-              fontFamily: 'var(--font-geist-mono, monospace)',
-              fontSize: 12,
-            }}
-          >
-            Loading playbooks…
-          </div>
-        ) : filtered.length === 0 ? (
-          <div
-            className="pg-panel"
-            style={{
-              padding: 24,
-              textAlign: 'center',
-              color: 'var(--fg-3)',
-              fontFamily: 'var(--font-geist-mono, monospace)',
-              fontSize: 12,
-            }}
-          >
-            No playbooks match this filter.
-          </div>
-        ) : view === 'grid' ? (
-          <div
-            style={{
-              display: 'grid',
-              gridTemplateColumns: 'repeat(3, 1fr)',
-              gap: 12,
-              paddingBottom: 24,
-            }}
-          >
-            {filtered.map((p) => (
-              <PlaybookCard
-                key={String(p.template_id || p.id)}
-                playbook={p}
-                onRun={handleRun}
-              />
-            ))}
-          </div>
-        ) : (
-          <ListView items={filtered} onRun={handleRun} />
-        )}
-      </section>
+      <PlaybooksBody />
 
       <CreatePlaybookModal
         open={playbookOpen}
         onClose={() => setPlaybookOpen(false)}
       />
-    </div>
-  )
-}
-
-function ListView({
-  items,
-  onRun,
-}: {
-  items: PlaybookLike[]
-  onRun: (p: PlaybookLike) => void
-}) {
-  return (
-    <div className="pg-panel" style={{ padding: 0 }}>
-      <table className="act-table">
-        <thead>
-          <tr>
-            <th style={{ width: 38 }}></th>
-            <th>Playbook</th>
-            <th style={{ width: 160 }}>Trigger</th>
-            <th style={{ width: 70, textAlign: 'right' }}>Steps</th>
-            <th style={{ width: 80, textAlign: 'right' }}>Runs</th>
-            <th style={{ width: 90, textAlign: 'right' }}>Success</th>
-            <th style={{ width: 90 }}></th>
-          </tr>
-        </thead>
-        <tbody>
-          {items.map((p) => {
-            const trig = trigger(p)
-            const TrigIcon =
-              trig.kind === 'cron' ? Clock : trig.kind === 'webhook' ? TriggerZap : Hand
-            const runs = p.use_count ?? 0
-            const success = p.success_rate ?? null
-            return (
-              <tr
-                key={String(p.template_id || p.id)}
-                onClick={() => onRun(p)}
-              >
-                <td>
-                  <span
-                    style={{
-                      width: 24,
-                      height: 24,
-                      borderRadius: 5,
-                      background: 'var(--bg-2)',
-                      color: 'var(--fg)',
-                      display: 'inline-flex',
-                      alignItems: 'center',
-                      justifyContent: 'center',
-                    }}
-                  >
-                    <BookMarked style={{ width: 12, height: 12 }} />
-                  </span>
-                </td>
-                <td>
-                  <div className="nm">{p.name || 'Untitled'}</div>
-                  {p.description && (
-                    <div
-                      style={{
-                        fontSize: 10.5,
-                        color: 'var(--fg-3)',
-                        marginTop: 2,
-                        overflow: 'hidden',
-                        textOverflow: 'ellipsis',
-                        whiteSpace: 'nowrap',
-                        maxWidth: 460,
-                      }}
-                    >
-                      {p.description}
-                    </div>
-                  )}
-                </td>
-                <td>
-                  <span className="trigger">
-                    <TrigIcon style={{ width: 10, height: 10 }} />
-                    {trig.label}
-                  </span>
-                </td>
-                <td className="mono" style={{ textAlign: 'right' }}>
-                  {p.steps?.length ?? 0}
-                </td>
-                <td className="mono" style={{ textAlign: 'right' }}>
-                  {runs}
-                </td>
-                <td
-                  className="mono"
-                  style={{
-                    textAlign: 'right',
-                    color:
-                      success == null
-                        ? 'var(--fg-3)'
-                        : success >= 99
-                        ? 'hsl(82 50% 22%)'
-                        : success >= 95
-                        ? 'hsl(38 78% 27%)'
-                        : 'hsl(var(--accent))',
-                  }}
-                >
-                  {success == null ? '—' : `${Math.round(success)}%`}
-                </td>
-                <td>
-                  <button
-                    type="button"
-                    className="l-btn l-btn--p"
-                    style={{ height: 26, fontSize: 11.5, padding: '0 12px' }}
-                    onClick={(e) => {
-                      e.stopPropagation()
-                      onRun(p)
-                    }}
-                  >
-                    Run
-                  </button>
-                </td>
-              </tr>
-            )
-          })}
-        </tbody>
-      </table>
     </div>
   )
 }
